refactor: coloca todas as constantes em um único arquivo
