+++ nutrixpert/db.py
@@ -2,12 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from dotenv import load_dotenv
-import os
-
-load_dotenv()
-
-DATABASE_URL = os.getenv("DATABASE_URL")
+from nutrixpert.core.utils.constants import DATABASE_URL
 
 # Criação do engine
 engine = create_engine(DATABASE_URL)
